requirements.txt
# Core data science libraries
pandas>=2.0.0 # core data manipulation library
numpy>=1.24.0 # core numerical computing library

# Financial data
yfinance>=0.2.0 

# Market calendar utilities
pandas-market-calendars>=4.3.0

# Technical analysis indicators
ta>=0.11.0

# Configuration and HTTP
PyYAML>=6.0 # library for loading configuration files
requests>=2.31.0 # library for making HTTP requests

